Conflicts:
	WebContent/WEB-INF/springconfig/applicationContext.xml
	WebContent/WEB-INF/tiles/definitions/tile-definition.xml

+++ src/main/java/com/mum/edu/happycart/repository/SubCategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.mum.edu.happycart.domain.SubCategory;
 
-public interface SubCategoryRepository extends CrudRepository<SubCategory, Long> {
+public interface SubCategoryRepository extends CrudRepository<SubCategory, Integer> {
 
 }
